+++ buildSrc/src/main/kotlin/Dependencies.kt
@@ -12,11 +12,31 @@ object Dependencies : BuiltInDependencies() {
         provider = DependencyProviders.maven("https://kotlin.bintray.com/kotlinx") //TODO move it to kamp
     )
 
-    val githubApi = RawDependency("org.kohsuke", "github-api", { Versions.githubApi }, DependencyProviders.mavenCentral)(jvm())
+    val cli = kotlinx.dependency(
+        name = "kotlinx-cli",
+        version = { Versions.cli },
+        target = jvm("jvm"),
+        provider = DependencyProviders.maven("https://dl.bintray.com/kotlin/kotlin-dev")
+    )
+
+    val githubApi = RawDependency(
+        group = "org.kohsuke",
+        name = "github-api",
+        version = { Versions.githubApi },
+        provider = DependencyProviders.mavenCentral
+    )(jvm())
+
     val kotlinShell = RawDependency(
-        "eu.jrie.jetbrains",
-        "kotlin-shell-core",
-        { Versions.kotlinShell },
-        DependencyProviders.maven("https://dl.bintray.com/jakubriegel/kotlin-shell")
+        group = "eu.jrie.jetbrains",
+        name = "kotlin-shell-core",
+        version = { Versions.kotlinShell },
+        provider = DependencyProviders.maven("https://dl.bintray.com/jakubriegel/kotlin-shell")
+    )(jvm())
+
+    val dotenv = RawDependency(
+        group = "io.github.cdimascio",
+        name = "java-dotenv",
+        version = { Versions.dotenv },
+        provider = DependencyProviders.mavenCentral
     )(jvm())
 }