+++ buildSrc/src/main/kotlin/Versions.kt
@@ -5,6 +5,8 @@ object Versions {
     )
     const val tdLatest = "1.5.1"
     const val immutableCollections = "0.3"
+    const val cli = "0.2.0-dev-6"
     const val githubApi = "1.99"
     const val kotlinShell = "0.2"
+    const val dotenv = "5.1.3"
 }