+++ cli/src/main/kotlin/dev/whyoleg/ktd/cli/api/builder/AdditionBuilder.kt
@@ -1,0 +1,13 @@
+package dev.whyoleg.ktd.cli.api.builder
+
+import dev.whyoleg.ktd.cli.api.tl.*
+
+fun StringBuilder.buildAnnotations(additions: List<TlAddition>) {
+    additions
+        .filterIsInstance<TlAddition.Annotation>()
+        .takeIf(List<*>::isNotEmpty)
+        ?.map(TlAddition.Annotation::annotation)
+        ?.distinct()
+        ?.sorted()
+        ?.joinTo(this, "\n@", "@", "\n")
+}