+++ cli/src/main/kotlin/dev/whyoleg/ktd/cli/api/builder/Conversion.kt
@@ -1,0 +1,3 @@
+package dev.whyoleg.ktd.cli.api.builder
+
+fun String.snakeToCamel(): String = split("_").joinToString("") { it.capitalize() }.decapitalize()