+++ cli/src/main/kotlin/dev/whyoleg/ktd/cli/api/tl/RawTlData.kt
@@ -1,0 +1,5 @@
+package dev.whyoleg.ktd.cli.api.tl
+
+data class RawTlData(val type: RawTlDataType, val list: List<String>, val expression: String)
+
+enum class RawTlDataType { Object, Function }