+++ cli/src/main/kotlin/dev/whyoleg/ktd/cli/api/tl/TlProperty.kt
@@ -1,0 +1,9 @@
+package dev.whyoleg.ktd.cli.api.tl
+
+data class TlProperty(
+    val name: String,
+    val type: TlType,
+    val descriptions: List<String>,
+    val additions: List<TlAddition>
+)
+