+++ generator/src/main/kotlin/dev/whyoleg/ktd/generator/api/builder/Conversion.kt
@@ -1,3 +1,0 @@
-package dev.whyoleg.ktd.generator.api.builder
-
-fun String.snakeToCamel(): String = split("_").joinToString("") { it.capitalize() }.decapitalize()