+++ generator/src/main/kotlin/dev/whyoleg/ktd/generator/api/tl/TlMetadata.kt
@@ -1,7 +1,0 @@
-package dev.whyoleg.ktd.generator.api.tl
-
-data class TlMetadata(
-    val descriptions: List<String>,
-    val additions: List<TlAddition>,
-    val properties: List<TlProperty>
-)